refactor: move custom ColumnType implementations to separate directory
